+++ scripts/BuildWinX64.ps1
@@ -1,7 +1,7 @@
 param()
 
 [string] $Product = "Mzinga"
-[string] $Target = "Win64"
+[string] $Target = "WinX64"
 
 & "$PSScriptRoot\Build.ps1" -Product $Product -Target $Target -BuildArgs "-target:Publish -p:RuntimeIdentifier=win-x64 -p:PublishSingleFile=true -p:PublishTrimmed=true -p:IncludeAllContentForSelfExtract=true -p:EnableCompressionInSingleFile=true"